LayoutTests/platform/ios-sim-deprecated/media/video-no-autoplay-expected.txt

Test that play event does not fire when "src" set with no autoplay attribute.

EXPECTED (video.paused == 'true') OK
EVENT(canplaythrough)
EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
END OF TEST
